WIP fixes Windows warnings

+++ src/chrono_granular/api/ChApiGranularChrono.cpp
@@ -25,7 +25,7 @@ void ChGranularChronoTriMeshAPI::load_meshes(std::vector<std::string> objfilenam
                                              std::vector<chrono::ChMatrix33<float>> rotscale,
                                              std::vector<float3> translations,
                                              std::vector<float> masses) {
-    unsigned int size = objfilenames.size();
+    unsigned int size = (unsigned int)objfilenames.size();
     if (size != rotscale.size() || size != translations.size() || size != masses.size()) {
         GRANULAR_ERROR("Mesh loading vectors must all have same size\n");
     }
@@ -96,9 +96,9 @@ void ChGranularChronoTriMeshAPI::setupTriMesh(const std::vector<chrono::geometry
         for (int i = 0; i < n_triangles_mesh; i++) {
             chrono::geometry::ChTriangle tri = mesh.getTriangle(i);
 
-            pMeshSoup->node1[tri_i] = make_float3(tri.p1.x(), tri.p1.y(), tri.p1.z());
-            pMeshSoup->node2[tri_i] = make_float3(tri.p2.x(), tri.p2.y(), tri.p2.z());
-            pMeshSoup->node3[tri_i] = make_float3(tri.p3.x(), tri.p3.y(), tri.p3.z());
+            pMeshSoup->node1[tri_i] = make_float3((float)tri.p1.x(), (float)tri.p1.y(), (float)tri.p1.z());
+            pMeshSoup->node2[tri_i] = make_float3((float)tri.p2.x(), (float)tri.p2.y(), (float)tri.p2.z());
+            pMeshSoup->node3[tri_i] = make_float3((float)tri.p3.x(), (float)tri.p3.y(), (float)tri.p3.z());
 
             pMeshSoup->triangleFamily_ID[tri_i] = family;
